% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageT.R
\name{averageT}
\alias{averageT}
\title{Averages out the sufficient statistics T(y)}
\usage{
averageT(y)
}
\arguments{
\item{y}{vector of values}
}
\value{
A vector with the average of the sufficient statistics
}
\description{
Takes a vector of values and returns the column average of their
sufficient statistic (determined by their class)
}
