% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericVector.R
\name{extractParameters}
\alias{extractParameters}
\title{Extract parameters}
\usage{
extractParameters(y, family = "gaussian", natural = FALSE, ...)
}
\arguments{
\item{y}{Numeric vector containing observations from a random variable}

\item{family}{Distribution family to assume for \code{y}}

\item{natural}{Should output be in terms of the natural parameter eta?}

\item{...}{arguments passed to \code{\link[=init.parms]{init.parms()}}}
}
\description{
Extract parameters
}
\examples{
x <- rtrunc(n = 100, lambda = 2, family = "Poisson")
attributes(x) <- NULL
class(x)
extractParameters(x, family = "normal")
extractParameters(x, family = "normal", natural = TRUE)
extractParameters(x, family = "binomial", nsize = 3)
extractParameters(x, family = "poisson", natural = FALSE)
extractParameters(x, family = "poisson", natural = TRUE)
}
