% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genRtruncClass.R
\name{genRtruncClass}
\alias{genRtruncClass}
\title{Generates an rtrunc-dispatchable class}
\usage{
genRtruncClass(n, family, parms)
}
\arguments{
\item{...}{parameter list passed to \code{rtrunc}}
}
\value{
A character string.
}
\description{
Matches a list of arguments to an rtrunc method
}
\author{
Waldir Leoncio
}
