% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{parameters2natural.beta}
\alias{parameters2natural.beta}
\title{Convert parameters to Natural parameters}
\usage{
parameters2natural.beta(parms)
}
\arguments{
\item{parms}{The shape1 and shape2 parameters in a beta distribution}
}
\value{
The natural parameters
}
\description{
Convert parameters to Natural parameters
}
\examples{
sample.lognorm <- rtrunc(n = 100000, mulog = 2.5, sigmalog = 0.5, a = 7)
ml_lognormal <- ml.estimation.trunc.dist(
  sample.lognorm, y.min = 7, max.it = 500, tol = 1e-10, delta = 0.3,
  family = "LogNormal"
)
eta.hat <- parameters2natural.beta(ml_lognormal)
}
\author{
René Holst
}
