% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.r
\name{parameters2natural.gamma}
\alias{parameters2natural.gamma}
\title{Convert parameters to Natural Gamma}
\usage{
parameters2natural.gamma(parms)
}
\arguments{
\item{parms}{The parameters alpha and beta in a gamma distribution}
}
\value{
The natural parameters
}
\description{
Convert parameters to Natural Gamma
}
\examples{
sample.lognorm <- rtrunc(n = 100000, mulog = 2.5, sigmalog = 0.5, a = 7)@sample
ml_lognormal <- ml.estimation.trunc.dist(
  sample.lognorm, y.min = 7, max.it = 500, tol = 1e-10, delta = 0.3,
  family = "LogNormal"
)
eta.hat <- parameters2natural.gamma(ml_lognormal)
}
\author{
René Holst
}
