% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{parameters2natural}
\alias{parameters2natural}
\title{Convert distribution parameters to natural parameters}
\usage{
parameters2natural(parms)
}
\arguments{
\item{parms}{A vector of parameters in a distribution distribution}
}
\value{
The natural parameters
}
\description{
Convert distribution parameters to natural parameters
}
\examples{
# Poisson distribution
samp <- rtrunc(n=100, lambda=2, family="Poisson")
TruncExpFam:::parameters2natural(TruncExpFam:::init.parms(samp))

}
