% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{parameters2natural.trunc_beta}
\alias{parameters2natural.trunc_beta}
\title{Convert parameters to Natural parameters}
\usage{
parameters2natural.trunc_beta(parms)
}
\arguments{
\item{parms}{The shape1 and shape2 parameters in a beta distribution}
}
\value{
The natural parameters
}
\description{
Convert parameters to Natural parameters
}
\examples{
sample.lognorm <- rtrunc(n=100000, meanlog=2.5, sdlog=0.5, a=7, family="log-normal")
ml_lognormal <- ml.estimation.trunc.dist(
  sample.lognorm, y.min = 7, max.it = 500, tol = 1e-10, delta = 0.3
)
eta.hat <- parameters2natural.beta(ml_lognormal)
}
\author{
René Holst
}
