% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{parameters2natural.trunc_gamma}
\alias{parameters2natural.trunc_gamma}
\title{Convert parameters to Natural Gamma}
\usage{
parameters2natural.trunc_gamma(parms)
}
\arguments{
\item{parms}{The parameters shape and rate in a gamma distribution}
}
\value{
The natural parameters
}
\description{
Convert parameters to Natural Gamma
}
\examples{
sample.lognorm <- rtrunc(n = 100000, meanlog = 2.5, sdlog = 0.5, a = 7, family="log-normal")
ml_lognormal <- ml.estimation.trunc.dist(
  sample.lognorm, y.min = 7, max.it = 500, tol = 1e-10, delta = 0.3,
)
eta.hat <- parameters2natural.trunc_gamma(ml_lognormal)
}
\author{
René Holst
}
