% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.trunc}
\alias{print.trunc}
\alias{print.trunc_beta}
\alias{print.trunc_binomial}
\alias{print.trunc_chisq}
\alias{print.trunc_contbern}
\alias{print.trunc_exp}
\alias{print.trunc_gamma}
\alias{print.trunc_invgamma}
\alias{print.trunc_invgauss}
\alias{print.trunc_lognormal}
\alias{print.trunc_nbinom}
\alias{print.trunc_poisson}
\alias{print.trunc_normal}
\title{Print sample from truncated distribution}
\usage{
\method{print}{trunc}(x, details = FALSE, ...)
}
\arguments{
\item{x}{object to print}

\item{details}{if \code{FALSE} (default), hides the attributes of \code{x}}

\item{...}{other arguments passed to \code{\link[base:print.default]{base::print.default()}}}
}
\value{
\code{x} with or without its attributes
}
\description{
Special printing methods for trunc_* classes.
}
\author{
Waldir Leoncio
}
