% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probdist-class.R
\docType{class}
\name{probdist-class}
\alias{probdist-class}
\alias{probdist}
\title{Probability distribution class}
\value{
An RC class containing statistical properties of that distribution,
namely its name, parameter names and values and natural parameter names and
values.
}
\description{
An R object describing the properties of a probability
distribution.
}

\examples{
probdist(shape = 2, scale = .25, family = "gamma")
}
\author{
Waldir Leoncio
}
