% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtrunc.R
\name{qtrunc}
\alias{qtrunc}
\title{Quantile Function}
\usage{
qtrunc(p, family, ..., lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of quantiles}

\item{family}{distribution family to use}

\item{...}{\emph{named} distribution parameters and/or truncation limits
(\code{a}, \code{b})}

\item{lower.tail}{logical; if \code{TRUE}, probabilities are
\eqn{P(X <= x)}{P(X \leq x)} otherwise, \eqn{P(X > x)}}

\item{log.p}{logical; if \code{TRUE}, probabilities p are given as \code{log(p)}}
}
\value{
The quantile of \code{p}.
}
\description{
Calculates quantile for a given truncated distribution and
probability.
}
\examples{
qtrunc(0)
qtrunc(.6, family = "gaussian", mean = 5, sd = 10, b = 7)
qnorm(.6, mean = 5, sd = 10) # for comparison
}
