% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrunc.R
\name{rtrunc,numeric,numeric,ANY,missing,missing,numeric,numeric,missing-method}
\alias{rtrunc,numeric,numeric,ANY,missing,missing,numeric,numeric,missing-method}
\title{Method containing the parameters for the truncated gamma distribution}
\usage{
\S4method{rtrunc}{numeric,numeric,ANY,missing,missing,numeric,numeric,missing}(n, a, b, alpha, beta)
}
\arguments{
\item{n}{sample size}

\item{a}{point of left truncation}

\item{b}{point of right truncation}

\item{alpha}{shape of "parent" distribution}

\item{beta}{rate of "parent" distribution}
}
\value{
A sample of size n drawn from a truncated distribution
}
\description{
Random generation for the truncated exponential family distributions.
}
\examples{
x <- rtrunc(n=1000, prob=0.6, trials=20, a=4, b=10) # Binomial
x # whole object
sample.binom <- x@sample # sample (probably smaller than 15 due to a and b)
plot(table(sample.binom), ylab="Frequency", main="Freq. of sampled values")
}
\author{
René Holst, Waldir Leôncio
}
