% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrunc.R
\name{rtrunc,numeric,numeric,ANY,missing,missing,numeric,numeric,missing,missing,missing,missing-method}
\alias{rtrunc,numeric,numeric,ANY,missing,missing,numeric,numeric,missing,missing,missing,missing-method}
\title{Random Truncated Gamma}
\usage{
\S4method{rtrunc}{numeric,numeric,ANY,missing,missing,numeric,numeric,missing,missing,missing,missing}(n, a, b, alpha, beta)
}
\arguments{
\item{n}{sample size}

\item{a}{point of left truncation}

\item{b}{point of right truncation}

\item{alpha}{shape of "parent" distribution}

\item{beta}{rate of "parent" distribution}
}
\value{
A sample of size n drawn from a truncated distribution
}
\description{
Random generation for the truncated exponential family distributions.
}
\examples{
# Truncated binomial distribution
x <- rtrunc(n=1000, prob=0.6, trials=20, a=4, b=10)
str(x) # whole object
sample.binom <- x@sample # sample (probably smaller than 15 due to a and b)
plot(table(sample.binom), ylab="Frequency", main="Freq. of sampled values")

# TODO: add example from gamma (see old commits)

# Truncated Log-Normal distribution
sample.lognorm <- rtrunc(n=100000, mulog=2.5, sigmalog=0.5, a=7)@sample

hist(
  sample.lognorm, nclass = 35, xlim = c(0, 60), freq = FALSE,
   ylim = c(0, 0.15)
)

# Normal distribution
sample.norm <- rtrunc(n=10000,mu=2,sigma=1.5,a=-1)@sample
hist(sample.norm, nclass = 25)
}
\author{
René Holst, Waldir Leôncio
}
