% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrunc.R
\name{rtrunc,numeric,numeric,numeric,missing,missing,missing,numeric,numeric-method}
\alias{rtrunc,numeric,numeric,numeric,missing,missing,missing,numeric,numeric-method}
\title{Random Truncated Binomial}
\usage{
\S4method{rtrunc}{numeric,numeric,numeric,missing,missing,missing,numeric,numeric}(n, trials, prob, a, b)
}
\arguments{
\item{n}{sample size}

\item{trials}{number of trials}

\item{a}{point of left truncation}

\item{b}{point of right truncation}

\item{probs}{probability of success on each trial}
}
\value{
A sample of size n drawn from a truncated binomial distribution
}
\description{
Random Truncated Binomial
}
\note{
The effective sample size is reduced due to truncation.
}
\examples{
rtrunc(n=15L, prob=.4, a=5, b=10, trials=10)
}
\author{
René Holst, Waldir Leôncio
}
