% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrunc.R
\name{rtrunc.binomial}
\alias{rtrunc.binomial}
\title{Random Truncated Binomial}
\usage{
rtrunc.binomial(n, size, prob, a, b)
}
\arguments{
\item{size}{number of size}

\item{prob}{probability of success on each trial}
}
\description{
Random Truncated Binomial
}
