% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.r
\name{rtrunc.gamma}
\alias{rtrunc.gamma}
\title{Random Truncated Gamma}
\usage{
rtrunc.gamma(n, alpha, beta, a, b)
}
\arguments{
\item{n}{sample size}

\item{alpha}{shape of "parent" distribution}

\item{beta}{rate of "parent" distribution}

\item{a}{point of left truncation}

\item{b}{point of right truncation}
}
\value{
A sample of size n drawn from a truncated gamma distribution
}
\description{
Random Truncated Gamma
}
\note{
The effective sample size is reduced due to truncation.
}
\examples{
sample.gamma <- rtrunc.gamma(n = 10000, alpha = 6, beta = 2, a = 2)
hist(sample.gamma,nclass=15)
}
\author{
René Holst
}
