% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-normal.r
\name{rtrunc.lognorm}
\alias{rtrunc.lognorm}
\title{Random Truncated Log-Normal}
\usage{
rtrunc.lognorm(n, mu, sigma, a, b)
}
\arguments{
\item{n}{sample size}

\item{mu}{mean of un-truncated distribution}

\item{sigma}{standard deviation of un-truncated distribution}

\item{a}{point of left truncation}

\item{b}{point of right truncation}
}
\value{
A sample of size n drawn from a truncated log-normal distribution
}
\description{
Random Truncated Log-Normal
}
\note{
The effective sample size is reduced due to truncation.
}
\examples{
sample.lognorm <- rtrunc.lognorm(n = 100000, mu = 2.5, sigma = 0.5, a = 7)
hist(
  sample.lognorm, nclass = 35, xlim = c(0, 60), freq = FALSE,
   ylim = c(0, 0.15)
)
}
\author{
René Holst
}
