% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.r
\name{rtrunc.norm}
\alias{rtrunc.norm}
\title{Random Truncated Normal}
\usage{
rtrunc.norm(n, mu, sigma, a, b)
}
\arguments{
\item{n}{sample size}

\item{mu}{mean of "parent" distribution}

\item{sigma}{standard deviation of "parent" distribution}

\item{a}{point of left truncation}

\item{b}{point of right truncation}
}
\value{
A sample of size n drawn from a truncated normal distribution
}
\description{
Random Truncated Normal
}
\note{
The effective sample size is reduced due to truncation.
}
\examples{
sample.norm <- rtrunc.norm(n=10000,mu=2,sigma=1.5,a=-1)
hist(sample.norm, nclass = 25)
}
\author{
René Holst
}
