% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrunc.R
\name{rtrunc.pois}
\alias{rtrunc.pois}
\title{Random Truncated Poisson}
\usage{
rtrunc.pois(n, lambda, a, b)
}
\arguments{
\item{n}{sample size}

\item{lambda}{mean and var of "parent" distribution}

\item{a}{point of left truncation}

\item{b}{point of right truncation}
}
\value{
A sample of size n drawn from a truncated Poisson distribution
}
\description{
Random Truncated Poisson
}
\note{
The effective sample size is reduced due to truncation. a, and b are included in the domain
}
\examples{
sample.pois <- rtrunc.pois(1000, 10, 4)
hist(sample.pois)
}
\author{
René Holst
}
