% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruncExpFam.R
\docType{package}
\name{TruncExpFam}
\alias{TruncExpFam}
\title{Truncated Exponential Family}
\description{
TruncExpFam is an R package to handle truncated members from the
exponential family.
}
\note{
Found a bug? Want to suggest a feature? Contribute to the scientific
and open source communities by opening an issue on our home page.
Check the "BugReports" field on \code{packageDescription("TruncExpFam")} for
the URL.
}
\section{Supported distributions}{

\itemize{
\item Beta
\item Binomial
\item Chi-Square
\item Continuous Bernoulli
\item Exponential
\item Gamma
\item Inverse Gamma
\item Inverse Gaussian
\item Log-normal
\item Negative Binomial
\item Normal
\item Poisson
}
}

