% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_rtrunc.R
\name{validateFamilyParms}
\alias{validateFamilyParms}
\title{Validate family parameters}
\usage{
validateFamilyParms(family, parms, verbose = FALSE)
}
\arguments{
\item{family}{character with family distribution name}

\item{parms}{character vector with distribution parameter names}

\item{verbose}{print intermediate messages?}
}
\value{
list telling if family-parm combo is valid + the family name
}
\description{
Checks if a combination of distribution family and parameters is
valid.
}
\author{
Waldir Leoncio
}
