% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R, R/binomial.R, R/chisq.R,
%   R/contbernoulli.R, R/dtrunc.R
\name{dtruncbeta}
\alias{dtruncbeta}
\alias{dtruncbinom}
\alias{dtruncchisq}
\alias{dtrunccontbernoulli}
\alias{dtrunc}
\title{Probability Density Function}
\usage{
dtruncbeta(y, eta, a, b)

dtruncbinom(y, eta, a = 0, b, ...)

dtruncchisq(y, eta, a = 0, b)

dtrunccontbernoulli(y, eta, a = 0, b)

dtrunc(y, eta, a, b)
}
\arguments{
\item{y}{output from rtrunc}

\item{eta}{Natural parameters}

\item{a}{lower truncation limit}

\item{b}{upper truncation limit}

\item{...}{size}
}
\description{
Calculates the PDF for a given truncated distribution
}
\examples{
y <- rtrunc(50, mean = 5, sd = 2)
dtrunc(y, eta = c(0, -1))
}
