% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrunc.R
\name{dtrunc}
\alias{dtrunc}
\title{Probability Density Function}
\usage{
dtrunc(y, ...)
}
\arguments{
\item{y}{output from rtrunc}

\item{...}{arguments to be passed to the methods}
}
\description{
Calculates the PDF for a given truncated distribution
}
