% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dtrunc}
\alias{dtrunc}
\title{Probability Density Function}
\usage{
dtrunc(y, ...)
}
\arguments{
\item{y}{output from rtrunc}
}
\value{

}
\description{
Calculates the PDF for a given
}
\author{
Waldir Leoncio
}
