% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{dtrunc}
\alias{dtrunc}
\title{Probability Density Function}
\usage{
dtrunc(y, eta, a, b)
}
\arguments{
\item{y}{output from rtrunc}

\item{...}{arguments to be passed to the methods}
}
\description{
Calculates the PDF for a given truncated distribution
}
\examples{
y <- rtrunc(50, mean=5, sd=2)
dtrunc(y, eta=c(0, -1))

}
