% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genRtruncClass.R
\name{genRtruncClass}
\alias{genRtruncClass}
\title{Generates an rtrunc-dispatchable class}
\usage{
genRtruncClass(n, family, parms)
}
\arguments{
\item{n}{sample size}

\item{family}{distribution family}

\item{parms}{list of parameters passed to rtrunc (through the \code{...} element)}
}
\value{
A character string.
}
\description{
Matches a list of arguments to an rtrunc method
}
\author{
Waldir Leoncio
}
