% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{natural2parameters}
\alias{natural2parameters}
\title{Convert natural parameters to distribution parameters}
\usage{
natural2parameters(eta)
}
\arguments{
\item{eta}{vector of natural parameters}
}
\value{
Distribution parameters
}
\description{
Convert natural parameters to distribution parameters
}
\examples{
samp <- rtrunc(n = 100, lambda = 2, family = "Poisson")
TruncExpFam:::natural2parameters(TruncExpFam:::init.parms(samp))
}
