% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{parameters2natural}
\alias{parameters2natural}
\title{Convert distribution parameters to natural parameters}
\usage{
parameters2natural(parms)
}
\arguments{
\item{parms}{A vector of parameters in a distribution distribution}
}
\value{
The natural parameters
}
\description{
Convert distribution parameters to natural parameters
}
\examples{
# Beta distribution
# samp <- rtrunc(n=100000, meanlog=2.5, sdlog=0.5, a=7, family="log-normal")
# ml <- ml.estimation.trunc.dist(
#   samp, y.min = 7, max.it = 500, tol = 1e-10, delta = 0.3
# )
# eta.hat <- parameters2natural.trunc_beta(ml_lognormal)
#' sample.lognorm <- rtrunc(n = 100000, meanlog = 2.5, sdlog = 0.5, a = 7, family="log-normal")

# Gamma distribution
# ml_lognormal <- ml.estimation.trunc.dist(
#  sample.lognorm, y.min = 7, max.it = 500, tol = 1e-10, delta = 0.3,
#)
# eta.hat <- parameters2natural.trunc_gamma(ml_lognormal)
}
