% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrunc.R
\name{ptrunc}
\alias{ptrunc}
\alias{ptrunc.normal}
\title{Cummulative Distribution Function}
\usage{
ptrunc(q, family = "gaussian", lower.tail = TRUE, ...)

\method{ptrunc}{normal}(
  q,
  family,
  lower.tail = TRUE,
  log.p = FALSE,
  mean = 0,
  sd = 1,
  a = -Inf,
  b = Inf,
  ...
)
}
\arguments{
\item{q}{vector of quantiles}

\item{family}{distribution family to use}

\item{lower.tail}{logical; if \code{TRUE}, probabilities are \eqn{P(X <= x)}{P(X \leq x)} otherwise, \eqn{P(X > x)}.}

\item{...}{additional parameters to be passed to the distribution function.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{mean}{mean of parent distribution}

\item{sd}{standard deviation is parent distribution}

\item{a}{point of left truncation}

\item{b}{point of right truncation}
}
\value{
The cummulative probability of y.
}
\description{
Calculates the cumulative probability for a given truncated distribution
}
\note{
Either the common or the natural parameters must be provided.
}
\examples{
ptrunc(0)
ptrunc(6, family = "gaussian", mean = 5, sd = 10, b = 7)
pnorm(6, mean = 5, sd = 10) # for comparison
}
