% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrunc.R
\name{ptrunc}
\alias{ptrunc}
\title{Cummulative Distribution Function}
\usage{
ptrunc(q, family, ..., lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantiles}

\item{family}{distribution family to use}

\item{...}{\emph{named} distribution parameters and/or truncation limits (\code{a}, \code{b})}

\item{lower.tail}{logical; if \code{TRUE}, probabilities are \eqn{P(X <= x)}{P(X \leq x)} otherwise, \eqn{P(X > x)}}

\item{log.p}{logical; if \code{TRUE}, probabilities p are given as \code{log(p)}}
}
\value{
The cummulative probability of y.
}
\description{
Calculates the cumulative probability for a given truncated distribution
}
\note{
Either the common or the natural parameters must be provided.
}
\examples{
ptrunc(0)
ptrunc(6, family = "gaussian", mean = 5, sd = 10, b = 7)
pnorm(6, mean = 5, sd = 10) # for comparison
}
