% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrunc.R
\name{rtrunc.lognorm}
\alias{rtrunc.lognorm}
\title{Random Truncated Gamma}
\usage{
rtrunc.lognorm(n, mulog, sigmalog, a, b)
}
\arguments{
\item{n}{sample size}

\item{a}{point of left truncation}

\item{b}{point of right truncation}

\item{alpha}{shape of "parent" distribution}

\item{beta}{rate of "parent" distribution}

\item{mu}{mean of un-truncated distribution}

\item{sigma}{standard deviation of un-truncated distribution}
}
\value{
A sample of size n drawn from a truncated gamma distribution

A sample of size n drawn from a truncated log-normal distribution
}
\description{
Random Truncated Gamma

Random Truncated Log-Normal
}
\note{
The effective sample size is reduced due to truncation.

The effective sample size is reduced due to truncation.
}
\examples{
sample.lognorm <- rtrunc.lognorm(n = 100000, mu = 2.5, sigma = 0.5, a = 7)
hist(
  sample.lognorm, nclass = 35, xlim = c(0, 60), freq = FALSE,
   ylim = c(0, 0.15)
)
}
\author{
René Holst

René Holst
}
