% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrunc.R
\name{rtrunc}
\alias{rtrunc}
\title{The Truncated Exponential Family}
\usage{
rtrunc(n, a, b, trials, prob, alpha, beta, mu)
}
\arguments{
\item{n}{sample size}

\item{a}{point of left truncation}

\item{b}{point of right truncation}

\item{trials}{number of trials}

\item{prob}{probability of success on each trial}

\item{alpha}{shape of "parent" distribution}

\item{beta}{rate of "parent" distribution}

\item{mu}{mean of un-truncated distribution}
}
\value{
A sample of size n drawn from a truncated distribution
}
\description{
Random generation for the truncated exponential family distributions.
}
\note{
The effective sample size is reduced due to truncation.
}
\examples{
x <- rtrunc(n=1000, prob=0.6, trials=20, a=4, b=10) # Binomial
x # whole object
sample.binom <- x@sample # sample (probably smaller than 15 due to a and b)
plot(table(sample.binom), ylab="Frequency", main="Freq. of sampled values")
}
\author{
René Holst, Waldir Leôncio
}
