% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrunc.R
\name{rtrunc}
\alias{rtrunc}
\title{The Truncated Exponential Family}
\usage{
rtrunc(n, trials, prob, alfa, beta, mu, a, b)
}
\arguments{
\item{n}{sample size}

\item{trials}{number of trials}

\item{beta}{rate of "parent" distribution}

\item{a}{point of left truncation}

\item{b}{point of right truncation}

\item{probs}{probability of success on each trial}

\item{alpha}{shape of "parent" distribution}
}
\value{
A sample of size n drawn from a truncated distribution
}
\description{
Random generation for the truncated exponential family distributions.
}
\note{
The effective sample size is reduced due to truncation.
}
\examples{
x <- rtrunc(n=15, prob=.4, a=5, b=10, trials=10) # Binomial
x # whole object
x@sample # sample (probably smaller than 15 due to the a and b restrictions)
}
\author{
René Holst, Waldir Leôncio
}
