% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R, R/chisq.R, R/contbernoulli.R,
%   R/gamma.R, R/log-normal.R, R/normal.R, R/poisson.R, R/rtrunc.R
\name{rtrunc.binomial}
\alias{rtrunc.binomial}
\alias{rtrunc.chisq}
\alias{rtrunc.contbernoulli}
\alias{rtrunc.gamma}
\alias{rtrunc.lognormal}
\alias{rtrunc.normal}
\alias{rtrunc.poisson}
\alias{rtrunc}
\title{The Truncated Exponential Family}
\usage{
rtrunc.binomial(n, size, prob, a, b)

rtrunc.chisq(n, df, a, b)

rtrunc.contbernoulli(n, lambda, a, b)

rtrunc.gamma(n, shape, rate, a = 0, b = Inf)

rtrunc.lognormal(n, meanlog, sdlog, a, b)

rtrunc.normal(n, mean, sd, a, b)

rtrunc.poisson(n, lambda, a, b)

rtrunc(n, family = "gaussian", ...)
}
\arguments{
\item{n}{sample size}

\item{size}{number of size}

\item{prob}{probability of success on each trial}

\item{a}{point of left truncation}

\item{b}{point of right truncation}

\item{df}{degrees of freedom for "parent" distribution}

\item{lambda}{mean and var of "parent" distribution}

\item{shape}{shape of "parent" distribution}

\item{rate}{rate of "parent" distribution}

\item{meanlog}{mean of un-truncated distribution}

\item{sdlog}{standard deviation of un-truncated distribution}

\item{mean}{mean of parent distribution}

\item{sd}{standard deviation is parent distribution}

\item{family}{distribution family to use}

\item{...}{individual arguments to each distribution}
}
\value{
A sample of size n drawn from a truncated distribution
}
\description{
Random generation for the truncated exponential family distributions. Please ferer to the "Details" and "Examples" section for more information on how to use this function.
}
\details{
The best way to use this function is by calling the \code{rtrunc} generic with the \code{family} parameter of your choice. You can also specifically call one of the methods (e.g. \code{rtrunc.poisson(10, lambda=3)} instead of `rtrunc(10, family="poisson", lambda=3)), but the latter is more flexible (i.e., easily programmable) and more robust (i.e., it contains better error handling and validation procedures).
}
\note{
The effective sample size is reduced due to truncation.
}
\examples{
# Truncated binomial distribution
sample.binom <- rtrunc(1000, family="binomial", prob=0.6, size=20, a=4, b=10)
sample.binom
plot(table(sample.binom), ylab="Frequency", main="Freq. of sampled values")

# Truncated Log-Normal distribution
sample.lognorm <- rtrunc(
  n=100000, family="log-normal", meanlog=2.5, sdlog=0.5, a=7
)
summary(sample.lognorm)

hist(
  sample.lognorm, nclass = 35, xlim = c(0, 60), freq = FALSE,
  ylim = c(0, 0.15)
)

# Normal distribution
sample.norm <- rtrunc(n=10000, mean=2, sd=1.5, a=-1)
head(sample.norm)
hist(sample.norm, nclass = 25)

# Gamma distribution
sample.gamma <- rtrunc(n=10000, family="gamma", shape=6, rate=2, a=2)
hist(sample.gamma, nclass = 15)

# Poisson distribution
sample.pois <- rtrunc(n=1000, family="poisson", lambda=10, a=4)
sample.pois
plot(table(sample.pois))
}
\author{
René Holst, Waldir Leôncio
}
