% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrunc.R
\name{rtrunc}
\alias{rtrunc}
\alias{rtruncbeta}
\alias{rtruncbinom}
\alias{rtruncchisq}
\alias{rtrunccontbernoulli}
\alias{rtruncexp}
\alias{rtruncgamma}
\alias{rtruncinvgamma}
\alias{rtruncinvgauss}
\alias{rtrunclnorm}
\alias{rtruncnbinom}
\alias{rtruncnorm}
\alias{rtruncpois}
\title{The Truncated Exponential Family}
\usage{
rtrunc(n, family = "gaussian", ...)

rtruncbeta(n, shape1, shape2, a, b)

rtruncbinom(n, size, prob, a, b)

rtruncchisq(n, df, a, b)

rtrunccontbernoulli(n, lambda, a, b = 1)

rtruncexp(n, rate = 1, a, b)

rtruncgamma(n, shape, rate = 1, a = 0, b = Inf)

rtruncinvgamma(n, shape, rate = 1, a = 0, b = Inf)

rtruncinvgauss(n, m, s, a = 0, b = Inf)

rtrunclnorm(n, meanlog, sdlog, a, b)

rtruncnbinom(n, size, prob, mu, a, b = Inf)

rtruncnorm(n, mean, sd, a, b)

rtruncpois(n, lambda, a, b)
}
\arguments{
\item{n}{sample size}

\item{family}{distribution family to use}

\item{...}{individual arguments to each distribution}

\item{shape1}{positive shape parameter alpha}

\item{shape2}{positive shape parameter beta}

\item{a}{point of left truncation}

\item{b}{point of right truncation}

\item{size}{target for number of successful trials,
or dispersion parameter (the shape parameter of the gamma mixing
distribution). Must be strictly positive, need not be integer.}

\item{prob}{probability of success on each trial}

\item{df}{degrees of freedom for "parent" distribution}

\item{lambda}{mean and var of "parent" distribution}

\item{rate}{inverse gamma rate parameter}

\item{shape}{inverse gamma shape parameter}

\item{m}{vector of means}

\item{s}{vector of dispersion parameters}

\item{meanlog}{mean of un-truncated distribution}

\item{sdlog}{standard deviation of un-truncated distribution}

\item{mu}{alternative parametrization via mean}

\item{mean}{mean of parent distribution}

\item{sd}{standard deviation is parent distribution}
}
\value{
A sample of size n drawn from a truncated distribution
}
\description{
Random generation for the truncated exponential family distributions. Please ferer to the "Details" and "Examples" section for more information on how to use this function.
}
\details{
The best way to use this function is by calling the \code{rtrunc} generic with the \code{family} parameter of your choice. You can also specifically call one of the methods (e.g. \code{rtrunc.poisson(10, lambda=3)} instead of `rtrunc(10, family="poisson", lambda=3)), but the latter is more flexible (i.e., easily programmable) and more robust (i.e., it contains better error handling and validation procedures).
}
\note{
The effective sample size is reduced due to truncation.
}
\examples{
# Truncated binomial distribution
sample.binom <- rtrunc(1000, family="binomial", prob=0.6, size=20, a=4, b=10)
sample.binom
plot(table(sample.binom), ylab="Frequency", main="Freq. of sampled values")

# Truncated Log-Normal distribution
sample.lognorm <- rtrunc(
  n=100000, family="lognormal", meanlog=2.5, sdlog=0.5, a=7
)
summary(sample.lognorm)

hist(
  sample.lognorm, nclass = 35, xlim = c(0, 60), freq = FALSE,
  ylim = c(0, 0.15)
)

# Normal distribution
sample.norm <- rtrunc(n=10000, mean=2, sd=1.5, a=-1)
head(sample.norm)
hist(sample.norm, nclass = 25)

# Gamma distribution
sample.gamma <- rtrunc(n=10000, family="gamma", shape=6, rate=2, a=2)
hist(sample.gamma, nclass = 15)

# Poisson distribution
sample.pois <- rtrunc(n=1000, family="poisson", lambda=10, a=4)
sample.pois
plot(table(sample.pois))
}
\author{
René Holst, Waldir Leôncio
}
