% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrunc.R
\name{rtrunc}
\alias{rtrunc}
\alias{rtrunc.gamma}
\alias{rtrunc.lognormal}
\alias{rtrunc.normal}
\alias{rtrunc.poisson}
\alias{rtrunc.contbernoulli}
\alias{rtrunc.chisq}
\title{The Truncated Exponential Family}
\usage{
rtrunc(n, family = "gaussian", ...)

rtrunc.gamma(n, shape, rate, a = 0, b = Inf)

rtrunc.lognormal(n, meanlog, sdlog, a, b)

rtrunc.normal(n, mean, sd, a, b)

rtrunc.poisson(n, lambda, a, b)

rtrunc.contbernoulli(n, lambda, a, b)

rtrunc.chisq(n, df, a, b)
}
\arguments{
\item{n}{sample size}

\item{shape}{shape of "parent" distribution}

\item{rate}{rate of "parent" distribution}

\item{a}{point of left truncation}

\item{b}{point of right truncation}

\item{meanlog}{mean of un-truncated distribution}

\item{sdlog}{standard deviation of un-truncated distribution}

\item{mean}{mean of parent distribution}

\item{sd}{standard deviation is parent distribution}

\item{lambda}{mean of "parent" distribution}

\item{df}{degrees of freedom for "parent" distribution}
}
\value{
A sample of size n drawn from a truncated distribution
}
\description{
Random generation for the truncated exponential family distributions.
}
\note{
The effective sample size is reduced due to truncation.
}
\examples{
# Truncated binomial distribution
sample.binom <- rtrunc(n=1000, prob=0.6, size=20, a=4, b=10)
sample.binom
plot(table(sample.binom), ylab="Frequency", main="Freq. of sampled values")

# Truncated Log-Normal distribution
sample.lognorm <- rtrunc(n=100000, meanlog=2.5, sdlog=0.5, a=7)
summary(sample.lognorm)

hist(
  sample.lognorm, nclass = 35, xlim = c(0, 60), freq = FALSE,
   ylim = c(0, 0.15)
)

# Normal distribution
sample.norm <- rtrunc(n=10000,mean=2,sd=1.5,a=-1)
head(sample.norm)
hist(sample.norm, nclass = 25)

# Gamma distribution
sample.gamma <- rtrunc(n = 10000, shape = 6, rate = 2, a = 2)
hist(sample.gamma, nclass = 15)

# Poisson distribution
sample.pois <- rtrunc(n=1000, lambda=10, a=4)
sample.pois
plot(table(sample.pois))
}
\author{
René Holst
}
