% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrunc.R
\name{rtrunc}
\alias{rtrunc}
\alias{rtrunc,numeric,numeric,missing,missing,missing,missing,missing-method}
\alias{rtrunc,missing,missing,numeric,missing,missing,missing,missing-method}
\alias{rtrunc,missing,missing,missing,numeric,missing,missing,missing-method}
\alias{rtrunc,missing,missing,missing,missing,numeric,missing,missing-method}
\alias{rtrunc,missing,missing,missing,missing,missing,numeric,missing-method}
\alias{rtrunc,numeric,missing,missing,missing,missing,missing,missing-method}
\alias{rtrunc,missing,missing,missing,missing,missing,missing,numeric-method}
\title{The Truncated Exponential Family}
\usage{
rtrunc(n, size, prob, alpha, beta, mulog, sigmalog, mean, sd, lambda, df, a, b)

\S4method{rtrunc}{numeric,numeric,missing,missing,missing,missing,missing}(n, size, prob, a, b)

\S4method{rtrunc}{missing,missing,numeric,missing,missing,missing,missing}(n, alpha, beta, a = 0, b = Inf)

\S4method{rtrunc}{missing,missing,missing,numeric,missing,missing,missing}(n, mulog, sigmalog, a, b)

\S4method{rtrunc}{missing,missing,missing,missing,numeric,missing,missing}(n, mean, sd, a, b)

\S4method{rtrunc}{missing,missing,missing,missing,missing,numeric,missing}(n, lambda, a, b)

\S4method{rtrunc}{numeric,missing,missing,missing,missing,missing,missing}(n, prob, a, b)

\S4method{rtrunc}{missing,missing,missing,missing,missing,missing,numeric}(n, df, a, b)
}
\arguments{
\item{n}{sample size}

\item{size}{number of size}

\item{prob}{mean of "parent" distribution}

\item{alpha}{shape of "parent" distribution}

\item{beta}{rate of "parent" distribution}

\item{mulog}{mean of un-truncated distribution}

\item{sigmalog}{standard deviation of un-truncated distribution}

\item{mean}{mean of parent distribution}

\item{sd}{standard deviation is parent distribution}

\item{lambda}{mean and var of "parent" distribution}

\item{df}{degrees of freedom for "parent" distribution}

\item{a}{point of left truncation}

\item{b}{point of right truncation}
}
\value{
A sample of size n drawn from a truncated distribution
}
\description{
Random generation for the truncated exponential family distributions.
}
\note{
The effective sample size is reduced due to truncation.
}
\examples{
# Truncated binomial distribution
sample.binom <- rtrunc(n=1000, prob=0.6, size=20, a=4, b=10)
sample.binom
plot(table(sample.binom), ylab="Frequency", main="Freq. of sampled values")

# Truncated Log-Normal distribution
sample.lognorm <- rtrunc(n=100000, mulog=2.5, sigmalog=0.5, a=7)
summary(sample.lognorm)

hist(
  sample.lognorm, nclass = 35, xlim = c(0, 60), freq = FALSE,
   ylim = c(0, 0.15)
)

# Normal distribution
sample.norm <- rtrunc(n=10000,mean=2,sd=1.5,a=-1)
head(sample.norm)
hist(sample.norm, nclass = 25)

# Gamma distribution
sample.gamma <- rtrunc(n = 10000, alpha = 6, beta = 2, a = 2)
hist(sample.gamma, nclass = 15)

# Poisson distribution
sample.pois <- rtrunc(n=1000, lambda=10, a=4)
sample.pois
plot(table(sample.pois))
}
\author{
René Holst
}
