% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrunc.R
\name{rtrunc}
\alias{rtrunc}
\title{The Truncated Exponential Family}
\usage{
rtrunc(n, a, b, prob, trials, alpha, beta, mulog, sigmalog, mu, sigma, lambda)
}
\arguments{
\item{n}{sample size}

\item{a}{point of left truncation}

\item{b}{point of right truncation}

\item{prob}{probability of success on each trial}

\item{trials}{number of trials}

\item{alpha}{shape of "parent" distribution}

\item{beta}{rate of "parent" distribution}

\item{mulog}{mean of un-truncated distribution}

\item{sigmalog}{standard deviation of un-truncated distribution}

\item{mu}{mean of parent distribution}

\item{sigma}{standard deviation is parent distribution}

\item{lambda}{mean and var of "parent" distribution}
}
\value{
A sample of size n drawn from a truncated distribution
}
\description{
Random generation for the truncated exponential family distributions.
}
\note{
The effective sample size is reduced due to truncation.
}
\examples{
# Truncated binomial distribution
x <- rtrunc(n=1000, prob=0.6, trials=20, a=4, b=10)
str(x) # whole object
sample.binom <- x@sample # sample (probably smaller than 15 due to a and b)
plot(table(sample.binom), ylab="Frequency", main="Freq. of sampled values")

# Truncated Log-Normal distribution
sample.lognorm <- rtrunc(n=100000, mulog=2.5, sigmalog=0.5, a=7)@sample

hist(
  sample.lognorm, nclass = 35, xlim = c(0, 60), freq = FALSE,
   ylim = c(0, 0.15)
)

# Normal distribution
sample.norm <- rtrunc(n=10000,mu=2,sigma=1.5,a=-1)@sample
hist(sample.norm, nclass = 25)

# Gamma distribution
sample.gamma <- rtrunc(n = 10000, alpha = 6, beta = 2, a = 2)@sample
hist(sample.gamma, nclass = 15)

# Poisson distribution
sample.pois <- rtrunc(n=1000, lambda=10, a=4)@sample
hist(sample.pois)
}
\author{
René Holst
}
